refactor: consolidate runUndo and runRedo into runNavigation
